requirements.txt
pandas
pandas-ta
python-telegram-bot
NorenRestApiPy
websocket-client
flask
werkzeug
requests